style: apply feedback from review
