+++ src/useSubscription.ts
@@ -1,20 +1,11 @@
 import { useEffect } from 'react';
 import { GraphQLSubscriptionConfig, requestSubscription, OperationType } from 'relay-runtime';
+import {
+    SkipGraphQLSubscriptionConfig,
+    SkipSubscriptionConfig,
+    SubscriptionConfig,
+} from './RelayHooksTypes';
 import { useRelayEnvironment } from './useRelayEnvironment';
-
-export type SubscriptionConfig = {
-    skip?: boolean;
-};
-
-export type SkipSubscriptionConfig = {
-    skip: true;
-};
-
-export interface SkipGraphQLSubscriptionConfig<TSubscription extends OperationType>
-    extends Omit<GraphQLSubscriptionConfig<TSubscription>, 'variables' | 'subscription'> {
-    subscription?: GraphQLSubscriptionConfig<TSubscription>['subscription'];
-    variables?: TSubscription['variables'];
-}
 
 export function useSubscription<TSubscriptionPayload extends OperationType = OperationType>(
     config: GraphQLSubscriptionConfig<TSubscriptionPayload>,
